+++ sync/SyncDataCollection.js
@@ -19,7 +19,7 @@ var Defer = require("../../ts-promises/Defer");
  * @since 2016-1-29
  */
 var SyncDataCollection = (function () {
-    /**
+    /** Create an object which can sync a data collection to/from a remote data destination/source
      * @param getLastSyncDownTimestamp a function to get a collection's last sync down timestamp (unix style millisecond number)
      * @param updateLastSyncDownTimestamp a function to update a collection's last sync down timestamp
      * @param isDeletedPropName the name of the property on both local and remote data models which contains the item's 'deleted' boolean flag
@@ -71,18 +71,33 @@ var SyncDataCollection = (function () {
         }
         function saveData() {
             // update the last sync time for this table to right now
-            self.updateLastSyncDownTimestamp(table);
+            try {
+                self.updateLastSyncDownTimestamp(table);
+            }
+            catch (err) {
+                syncFailure(err);
+            }
             dfd.resolve(null);
         }
-        syncDownFunc(params).done(function (items) {
-            var promise = processResultsCallback(items);
-            if (promise != null && promise["then"]) {
-                promise.then(saveData, syncFailure);
-            }
-            else {
-                saveData();
-            }
-        }, syncFailure);
+        try {
+            syncDownFunc(params).done(function (items) {
+                try {
+                    var promise = processResultsCallback(items);
+                    if (promise != null && promise["then"]) {
+                        promise.then(saveData, syncFailure);
+                    }
+                    else {
+                        saveData();
+                    }
+                }
+                catch (err) {
+                    syncFailure(err);
+                }
+            }, syncFailure);
+        }
+        catch (err) {
+            syncFailure(err);
+        }
         return dfd.promise;
     };
     /** Using the URL, data collection, and other settings from 'syncSettings', synchronize data from a local data collection to a remove server
@@ -99,13 +114,13 @@ var SyncDataCollection = (function () {
         var primaryKey = Arrays.getIfOneItem(primaryKeys);
         var localColl = syncSetting.localCollection;
         return this.syncAndUpdateCollection(localColl, copyItemFunc, primaryKey, primaryKeys, function convertAndSendItemsToServer(items) {
-            var toServiceConverter = syncSetting.convertToSvcObjectFunc;
+            var toSvcObj = syncSetting.convertToSvcObjectFunc;
             var data = null;
             if (primaryKey) {
-                data = SyncDataCollection.checkAndConvertSingleKeyItems(localColl.getName(), items, primaryKey, toServiceConverter);
+                data = SyncDataCollection.checkAndConvertSingleKeyItems(localColl.getName(), items, primaryKey, toSvcObj);
             }
             else {
-                data = SyncDataCollection.checkAndConvertMultiKeyItems(localColl.getName(), items, primaryKeys, toServiceConverter);
+                data = SyncDataCollection.checkAndConvertMultiKeyItems(localColl.getName(), items, primaryKeys, toSvcObj);
             }
             return syncSetting.syncUpFunc(params, data).then(function (res) {
                 return res;
@@ -149,55 +164,56 @@ var SyncDataCollection = (function () {
             else {
                 self.updateMultiPrimaryKeyItems(table, items, primaryKeys);
             }
-            dfd.resolve(null);
+            dfd.resolve(res);
         }, function (err) {
             dfd.reject(err);
         });
         return dfd.promise;
     };
-    /** Search for items in a table using multiple primary keys and 'lastModifiedPropName' and use updateWhere(...) to update matching items
+    /** Search for items in a table using multiple primary keys and 'isSynchedPropName' and use updateWhere(...) to update matching items
      */
     SyncDataCollection.prototype.updateMultiPrimaryKeyItems = function (table, items, primaryKeys) {
         // set each item's synched flag to true once the items have been synched with the server
         var synchedProp = {};
         synchedProp[this.isSynchedPropName] = true;
         var whereFilter = {};
-        whereFilter[this.lastModifiedPropName] = { $lte: null };
         whereFilter[this.isSynchedPropName] = false;
         for (var k = 0, keyCount = primaryKeys.length; k < keyCount; k++) {
             whereFilter[primaryKeys[k]] = null;
         }
         for (var i = 0; i < items.length; ++i) {
             var item = items[i];
-            whereFilter[this.lastModifiedPropName].$lte = item[this.lastModifiedPropName] + 50;
             for (var k = 0, keyCount = primaryKeys.length; k < keyCount; k++) {
                 whereFilter[primaryKeys[k]] = item[primaryKeys[k]];
             }
             table.updateWhere(whereFilter, synchedProp);
         }
     };
-    /** Search for items in a table using a single primary key and 'lastModifiedPropName' and use updateWhere(...) to update matching items
+    /** Search for items in a table using a single primary key and 'isSynchedPropName' and use updateWhere(...) to update matching items
      */
     SyncDataCollection.prototype.updateSinglePrimaryKeyItems = function (table, items, primaryKey) {
         // set each item's synched flag to true once the items have been synched with the server
         var synchedProp = {};
         synchedProp[this.isSynchedPropName] = true;
         var whereFilter = {};
-        whereFilter[this.lastModifiedPropName] = { $lte: null };
         whereFilter[this.isSynchedPropName] = false;
         whereFilter[primaryKey] = null;
         for (var i = 0; i < items.length; ++i) {
             var item = items[i];
-            whereFilter[this.lastModifiedPropName].$lte = item[this.lastModifiedPropName] + 50;
             whereFilter[primaryKey] = item[primaryKey];
             table.updateWhere(whereFilter, synchedProp);
         }
     };
-    /** remove items marked for deletion and add new items or update existing items depending on parameters
+    /** Create a function which removes items marked for deletion and add new items or update existing items depending on parameters.
+     * @param syncSettings the sync down settings required to create a sync down items function
+     * @param isDeletedPropName the name of the property which indicates if an item should be deleted, if value of this property is truthy the item is deleted when the syncing, nullable
+     * @param syncDownOp the type of merge/update/add operation to perform with the new items
      */
     SyncDataCollection.createAddUpdateOrRemoveItemsFunc = function (syncSettings, isDeletedPropName, syncDownOp) {
         return function addUpdateOrRemoveItemsFunc(items) {
             var table = syncSettings.localCollection;
+            var findFilterFunc = syncSettings.findFilterFunc;
+            var convertToLocalObjectFunc = syncSettings.convertToLocalObjectFunc;
             if (syncDownOp.removeAll) {
                 table.clearCollection();
             }
@@ -209,13 +225,13 @@ var SyncDataCollection = (function () {
                         var item = items[i];
                         if (isDeletedPropName != null && item[isDeletedPropName]) {
                             if (removeDeletedData) {
-                                var query = syncSettings.findFilterFunc(item);
+                                var query = findFilterFunc(item);
                                 table.removeWhere(query);
                             }
                         }
                         else {
-                            var convertedItem = syncSettings.convertToLocalObjectFunc(item);
-                            var query = syncSettings.findFilterFunc(item);
+                            var convertedItem = convertToLocalObjectFunc(item);
+                            var query = findFilterFunc(item);
                             table.addOrUpdateWhereNoModify(query, convertedItem);
                         }
                     }
@@ -226,12 +242,12 @@ var SyncDataCollection = (function () {
                         var item = items[i];
                         if (isDeletedPropName != null && item[isDeletedPropName]) {
                             if (removeDeletedData) {
-                                var query = syncSettings.findFilterFunc(item);
+                                var query = findFilterFunc(item);
                                 table.removeWhere(query);
                             }
                         }
                         else {
-                            var convertedItem = syncSettings.convertToLocalObjectFunc(item);
+                            var convertedItem = convertToLocalObjectFunc(item);
                             res.push(convertedItem);
                         }
                     }